requirement.txt
asgiref==3.4.1
Django==3.2.10
django-filter==21.1
gunicorn==20.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
